backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
google-auth==2.27.0
requests==2.31.0
motor==3.3.2
pymongo==4.6.1
email-validator>=2.1.1
openai>=2.8.0
python-docx==1.1.0
